+++ src/TimeProviderExtensions/AutoAdvanceBehavior.cs
@@ -8,15 +8,27 @@ namespace TimeProviderExtensions;
 public sealed record class AutoAdvanceBehavior
 {
     private TimeSpan clockAdvanceAmount = TimeSpan.Zero;
+    private TimeSpan timestampAdvanceAmount = TimeSpan.Zero;
 
     /// <summary>
-    /// Gets or sets the amount of time by which time advances whenever the clock is read via <see cref="TimeProvider.GetUtcNow"/> or <see cref="TimeProvider.GetLocalNow"/>.
+    /// Gets or sets the amount of time by which time advances whenever the clock is read via <see cref="TimeProvider.GetUtcNow"/>
+    /// or <see cref="TimeProvider.GetLocalNow"/>.
     /// </summary>
     /// <remarks>
     /// Set to <see cref="TimeSpan.Zero"/> to disable auto advance. The default value is <see cref="TimeSpan.Zero"/>.
     /// </remarks>
     /// <exception cref="ArgumentOutOfRangeException">Thrown when set to a value than <see cref="TimeSpan.Zero"/>.</exception>
     public TimeSpan ClockAdvanceAmount { get => clockAdvanceAmount; set { ThrowIfLessThanZero(value); clockAdvanceAmount = value; } }
+
+    /// <summary>
+    /// Gets or sets the amount of time by which time advances whenever the a timestamp is read via <see cref="TimeProvider.GetTimestamp"/>
+    /// or an elapsed time is calculated with <see cref="TimeProvider.GetElapsedTime(long)"/>.
+    /// </summary>
+    /// <remarks>
+    /// Set to <see cref="TimeSpan.Zero"/> to disable auto advance. The default value is <see cref="TimeSpan.Zero"/>.
+    /// </remarks>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown when set to a value than <see cref="TimeSpan.Zero"/>.</exception>
+    public TimeSpan TimestampAdvanceAmount { get => timestampAdvanceAmount; set { ThrowIfLessThanZero(value); timestampAdvanceAmount = value; } }
 
     private static void ThrowIfLessThanZero(TimeSpan value, [CallerMemberName] string? parameterName = null)
     {