+++ test/TimeProviderExtensions.Tests/AutoAdvanceBehaviorTests.cs
@@ -12,4 +12,15 @@ public class AutoAdvanceBehaviorTests
         throws.Should().Throw<ArgumentOutOfRangeException>()
             .And.ParamName.Should().Be(nameof(AutoAdvanceBehavior.ClockAdvanceAmount));
     }
+
+    [Fact]
+    public void TimestampAdvanceAmount_throws_when_lt_zero()
+    {
+        var sut = new AutoAdvanceBehavior();
+
+        var throws = () => sut.TimestampAdvanceAmount = TimeSpan.FromTicks(-1);
+
+        throws.Should().Throw<ArgumentOutOfRangeException>()
+            .And.ParamName.Should().Be(nameof(AutoAdvanceBehavior.TimestampAdvanceAmount));
+    }
 }